\name{BFFMod-package}
\alias{BFFMod-package}
\alias{BFFMod}
\docType{package}
\title{
  Physiological models of mussel growth in aquatic environments based on DEB theory
}
\description{
  Contains individual growth and population trajectory models based on dynamic energy budget.
  Models can be run in dynamic environments.
  
  The models are implemented in fortran and linked to R.
  
  Main functions: 
  
  \code{mussel_run_debind}
  
  \code{mussel_run_debcohort}
}

\author{
  Karline Soetaert and Quinten Mudde
}

\references{
model: 
Saraiva et al., 2011 

data on Mytilus galloprovincialis: 

Galimany et al., 2015 (physiological rates data) 

Karayusel et al., 2010 (Sinop growth data)  

Vural et al., 2015 (Marmara Sea growth data)  

Chelyadina et al., 2020 (Crimea growth data) 
}

\keyword{ package }
